Refactored out common tasks into governing task_group for TIGER DAGs.
