+++ airflow/great_expectations/great_expectations.yml
@@ -367,4 +367,16 @@ fluent_datasources:
         batch_metadata: {}
         table_name: united_states_counties_2022
         schema_name: data_raw
+      data_raw.temp_cook_county_areawaters_2022:
+        type: table
+        order_by: []
+        batch_metadata: {}
+        table_name: temp_cook_county_areawaters_2022
+        schema_name: data_raw
+      data_raw.cook_county_areawaters_2022:
+        type: table
+        order_by: []
+        batch_metadata: {}
+        table_name: cook_county_areawaters_2022
+        schema_name: data_raw
     connection_string: ${GX_DWH_DB_CONN}
